♻️Refector: Qr print function modularity
