+++ src/shared/types/name-tag/type.ts
@@ -1,0 +1,11 @@
+export interface TraineeData {
+  traineeId: number;
+  phoneNumber: string;
+}
+
+export interface UserData {
+  id: number;
+  name: string;
+  affiliation: string;
+  qrCode: string;
+}
